+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - [#73](https://github.com/Kashoo/synctos/issues/73): Include an implicit type property when a simple type filter is used
 - [#78](https://github.com/Kashoo/synctos/issues/78): Enum property validation type
+- [#79](https://github.com/Kashoo/synctos/issues/79): Support minimum/maximum size constraint on hashtable validation type
 
 ## [1.6.0] - 2017-01-18
 ### Added